Minor deployment tweaks for Heroku
Heroku was timing out when trying to
just deploy "npm start" because it was
thinking it was a web service, when it
is really a worker service.

This commit adds a Procfile to add that
 type of service for Heroku to use.

Also has some minor modifications of
the .gitignore file for style. I had to
make comment out some of the ignored
folders initially due to deployment
issues (lib & build), but that I don't
think is a real change.